pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "wgmesh"
description = "Wireguard Mesh Configuration Coordinator"
authors = [
    {name = "Joshua Schmidlkofer", email = "joshua+wgmesh@joshuainnovates.us" },
]
requires-python = ">=3.10"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Operating System :: Linux",
]
dependencies = [
    "typer[all]",
    "wheel",
    "attrs",
    "rich",
    "dnspython",
    "ifaddr",
    "loguru",
    "natsort",
    "netaddr",
    "pynacl",
    "route53",
    "ruamel.yaml",
    "six",
    "jinja2",
    "munch",
]
dynamic = ["version", "readme"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["wgmesh"]
namespaces = false

[tool.setuptools.dynamic]
version = {attr = "wgmesh.version.VERSION"}
readme = {file = ["README.md"], content-type = "text/markdown"}

[project.scripts]
wgsite   = "wgmesh.wgsite:app"
wgsetup  = "wgmesh.wgsetup:app"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

